src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --no-dev --no-emit-project -o src/requirements.txt
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
cloudevents==1.12.0
    # via functions-framework
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==46.0.5
    # via google-auth
deprecation==2.1.0
    # via cloudevents
feedgen==1.0.0
    # via google-alert-rss-proxy
feedparser==6.0.11
    # via google-alert-rss-proxy
flask==3.1.0
    # via
    #   functions-framework
    #   google-alert-rss-proxy
functions-framework==3.8.2
    # via google-alert-rss-proxy
google-api-core==2.25.2 ; python_full_version >= '3.14'
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-api-core==2.29.0 ; python_full_version < '3.14'
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-cloud-appengine-logging==1.8.0
    # via google-cloud-logging
google-cloud-audit-log==0.4.0
    # via google-cloud-logging
google-cloud-core==2.5.0
    # via google-cloud-logging
google-cloud-logging==3.11.3
    # via google-alert-rss-proxy
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-logging
grpcio==1.78.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==25.0.3 ; sys_platform != 'win32'
    # via functions-framework
idna==3.11
    # via requests
importlib-metadata==8.7.1
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
levenshtein==0.26.1
    # via python-levenshtein
lxml==6.0.2
    # via feedgen
markupsafe==3.0.3
    # via
    #   jinja2
    #   werkzeug
opentelemetry-api==1.39.1
    # via google-cloud-logging
packaging==26.0
    # via
    #   deprecation
    #   gunicorn
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==5.29.6
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
python-dateutil==2.9.0.post0
    # via feedgen
python-levenshtein==0.26.1
    # via google-alert-rss-proxy
rapidfuzz==3.14.3
    # via levenshtein
requests==2.32.3
    # via
    #   google-alert-rss-proxy
    #   google-api-core
rsa==4.9.1
    # via google-auth
sgmllib3k==1.0.0
    # via feedparser
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via
    #   grpcio
    #   opentelemetry-api
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.5
    # via
    #   flask
    #   functions-framework
zipp==3.23.0
    # via importlib-metadata
